Faster, less laggy and functional
The game items are now only drawn onto a surface that is as
big as the pygame window (or smaller), which reduced lag
when zooming in.
Started using sets instead of lists in Slitherlink which
also massivly reduced load time, but this will have to be
compensated when implementing rules.